requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
requests
groq
pandas
openpyxl
reportlab
